add vertex normals
So there's a weird thing.

rlLoadRenderBatch is always called with the default shader as the active shader.

This means that the binding of vertex data -- colors, positions, etc -- always
takes place based on the default shader's attribute locations, regardless of
the shader you're actually rendering.

This means that even if your custom shader has a vertexNormal attribute, it
won't be bound properly, because the *default* shader doesn't have one.

This is bad.

I don't really know how to fix this, so I didn't. Instead I added vertexNormal
attribute to the default vertex shader, which is ignored. I feel like this is
fragile, though, as I could imagine that attribute being optimized out since
it isn't actually used/necessary. And if it's optimized out of the default
shader, it's no longer available to *any* shader, even one that does use it.