fix MeshTangents
The implementation didn't take the indices buffer into account, so it only
worked with triangles, not quads. In the process of fixing this, I also
refactored the code so it's a little easier to follow.

But there was still a problem even with the correct tangent calculation: it
assumed that it was called with a dynamic mesh that had already been uploaded
to the GPU, and would segfault if called with a mesh that had not yet been
uploaded. So we check if we've been uploaded yet before we try to rebind the
attribute data.